feat(28): fixed 'task' interface
