+++ src/app/types/Task.ts
@@ -1,0 +1,7 @@
+export interface Task {
+  _id: string;
+  name: string;
+  tags: string[];
+  description: string;
+  weight: number;
+}
